+++ pkg/innerring/deploy.go
@@ -2,7 +2,6 @@ package innerring
 
 import (
 	"fmt"
-	"math"
 	"sync"
 
 	"github.com/nspcc-dev/neo-go/pkg/core/block"
@@ -13,8 +12,6 @@ import (
 	"github.com/nspcc-dev/neo-go/pkg/rpcclient/notary"
 	"github.com/nspcc-dev/neo-go/pkg/util"
 	"github.com/nspcc-dev/neofs-node/pkg/morph/client"
-	"github.com/nspcc-dev/neofs-node/pkg/morph/client/netmap"
-	"github.com/nspcc-dev/neofs-node/pkg/morph/deploy"
 )
 
 // [deploy.Blockchain] methods provided from both [client.Client] and
@@ -30,9 +27,6 @@ type neoFSSidechain struct {
 	neoFSSidechainCommonRPC
 
 	client *client.Client
-
-	netmapContractMtx sync.RWMutex
-	netmapContract    *netmap.Client
 
 	wsClient *rpcclient.WSClient
 
@@ -117,58 +111,4 @@ func newNeoFSSidechain(sidechainClient *client.Client, sidechainWSClient *rpccli
 	}
 
 	return res
-}
-
-func (x *neoFSSidechain) CurrentState() (deploy.NeoFSState, error) {
-	var res deploy.NeoFSState
-	var err error
-
-	x.netmapContractMtx.RLock()
-	netmapContract := x.netmapContract
-	x.netmapContractMtx.RUnlock()
-
-	if netmapContract == nil {
-		x.netmapContractMtx.Lock()
-
-		if x.netmapContract == nil {
-			netmapContractAddress, err := x.client.NNSContractAddress(client.NNSNetmapContractName)
-			if err != nil {
-				x.netmapContractMtx.Unlock()
-				return res, fmt.Errorf("resolve address of the '%s' contract in NNS: %w", client.NNSNetmapContractName, err)
-			}
-
-			x.netmapContract, err = netmap.NewFromMorph(x.client, netmapContractAddress, 0)
-			if err != nil {
-				x.netmapContractMtx.Unlock()
-				return res, fmt.Errorf("create Netmap contract client: %w", err)
-			}
-		}
-
-		netmapContract = x.netmapContract
-
-		x.netmapContractMtx.Unlock()
-	}
-
-	res.CurrentEpoch, err = netmapContract.Epoch()
-	if err != nil {
-		return res, fmt.Errorf("get current epoch from Netmap contract: %w", err)
-	}
-
-	res.CurrentEpochBlock, err = netmapContract.LastEpochBlock()
-	if err != nil {
-		return res, fmt.Errorf("get last epoch block from Netmap contract: %w", err)
-	}
-
-	epochDur, err := netmapContract.EpochDuration()
-	if err != nil {
-		return res, fmt.Errorf("get epoch duration from Netmap contract: %w", err)
-	}
-
-	if epochDur > math.MaxUint32 {
-		return res, fmt.Errorf("epoch duration from Netmap contract overflows uint32: %d", epochDur)
-	}
-
-	res.EpochDuration = uint32(epochDur)
-
-	return res, nil
 }