+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Changelog for NeoFS Node
 ### Fixed
 - Fund transfer deadlock in NeoFS chain auto-deploy/update procedure (#2681)
 - Invalid contracts' update transactions when epochs are stuck during the NeoFS chain update (#2680)
+- Metrics availability during startup (#2677) 
 
 ### Changed